requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pyyaml==6.0.2
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
openai==1.77.0
pydantic-settings==2.9.1
